Abstract class for sidebar factories
